Remove bucket.firestore and bucket.model and create a new model bucket
